+++ app/builders/dummy_builder.rb
@@ -4,4 +4,6 @@ class DummyBuilder
   def call(result:)
     result
   end
+
+  def equip(...); end
 end